Added MVC structure
