+++ models/user.js
@@ -1,0 +1,10 @@
+function CreateUser(name, room, session_id, profilePhoto){
+    return {
+        name,
+        room,
+        session_id,
+        profilePhoto
+    };
+}
+
+module.exports = { CreateUser };
