+++ views/main.html
@@ -67,7 +67,7 @@
         <div class="chat-container">
             <header class="chat-header">
                 <h1 class="main-heading"><i class="fas fa-smile"></i>GeekChat</h1>
-                <a href="index.html">
+                <a href="/">
                     <button class="foo-button mdc-button mdc-button--raised" style="background-color: red">
                         <div class="mdc-button__ripple"></div>
                         <span class="mdc-button__label">Leave</span>
